Add local storage handling for a service
This commit introduces two new files, 'storage/local.go' and 'storage/storage.go', to handle local storage for a service. This comprises the creation, deletion, retrieval, and editing of files as well as serving them on a route. This feature is added to improve storage handling in the service. Furthermore, slight changes to the 'http/routers/config.go' and 'go.sum' files were made to support this storage mechanism.